requirements.txt
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
certifi==2023.7.22
charset-normalizer==3.4.1
click==8.1.7
cssselect==1.2.0
fastapi==0.101.1
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
jmespath==1.0.1
lxml==4.9.3
mariadb==1.1.10
nest-asyncio==1.6.0
packaging==23.1
parsel==1.8.1
pybase64==1.3.0
pydantic==2.2.1
pydantic_core==2.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
PyYAML==6.0.1
requests==2.32.3
six==1.17.0
sniffio==1.3.0
SQLAlchemy==2.0.35
starlette==0.27.0
typing_extensions==4.7.1
urllib3==2.3.0
uvicorn==0.23.2
uvloop==0.17.0
w3lib==2.1.2
watchfiles==0.19.0
websockets==11.0.3
